# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DriveAssist API",
        "description": "API documentation for DriveAssist - Vehicle Diagnostics and Expert Marketplace",
        "contact": {
            "name": "DriveAssist Support",
            "email": "support@driveassist.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/quiz/questions/{category}": {
            "get": {
                "tags": [
                    "Road Signs"
                ],
                "summary": "Get quiz questions for category",
                "operationId": "9c0de47bfe42cea1e381772977f9507b",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/quiz/submit": {
            "post": {
                "tags": [
                    "Road Signs"
                ],
                "summary": "Submit quiz answers",
                "operationId": "c8362608adda1d053eea3585dc225fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {
                                        "type": "string"
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer"
                                                },
                                                "answer_index": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "time_taken": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/quiz/history": {
            "get": {
                "tags": [
                    "Road Signs"
                ],
                "summary": "Get quiz attempt history",
                "operationId": "37d8f2355f9f5f5435c4900e0abee902",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/road-signs/categories/{slug}": {
            "get": {
                "tags": [
                    "Road Signs"
                ],
                "summary": "Get signs by category",
                "operationId": "0abfbd20af395741f20da7d81b540bff",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/road-signs/{slug}": {
            "get": {
                "tags": [
                    "Road Signs"
                ],
                "summary": "Get road sign details",
                "operationId": "63947450a90c6e9576cc4ad6c4881776",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}